rename: ディレクトリ名はfeedのままではダメだ
